First commited as before_5764_rev1

+++ org.eclipse.jgit/src/org/eclipse/jgit/transport/TransportSftp.java
@@ -242,14 +242,13 @@ public class TransportSftp extends SshTransport implements WalkTransport {
 					if (!files.containsKey(in))
 						continue;
 
-					mtimes.put(n, Integer.valueOf(ent.getAttrs().getMTime()));
+					mtimes.put(n, ent.getAttrs().getMTime());
 					packs.add(n);
 				}
 
 				Collections.sort(packs, new Comparator<String>() {
 					public int compare(final String o1, final String o2) {
-						return mtimes.get(o2).intValue()
-								- mtimes.get(o1).intValue();
+						return mtimes.get(o2) - mtimes.get(o1);
 					}
 				});
 			} catch (SftpException je) {
